move layout into workspaceLayoutButton
